requirements.txt
dbt-core
dbt-postgres
sqlfluff[dbt]
lightdash-ops
pandas==2.2.1
psycopg==3.1.18
python-dotenv==1.0.1
openpyxl==3.1.2
tqdm==4.66.2
numpy==1.26.4 